Modified telemetry at the pass level; added casts to avoid warnings; added comments.

+++ src/inc/GlyphCache.h
@@ -149,14 +149,14 @@ float GlyphCache::getBoundingMetric(unsigned short glyphid, uint8 metric) const
 {
     if (glyphid >= _num_glyphs) return 0.;
     switch (metric) {
-        case 0: return glyph(glyphid)->theBBox().bl.x;                          // x_min
-        case 1: return glyph(glyphid)->theBBox().bl.y;                          // y_min
-        case 2: return glyph(glyphid)->theBBox().tr.x;                          // x_max
-        case 3: return glyph(glyphid)->theBBox().tr.y;                          // y_max
-        case 4: return _boxes[glyphid] ? _boxes[glyphid]->slant().bl.x : 0.;    // sum_min
-        case 5: return _boxes[glyphid] ? _boxes[glyphid]->slant().bl.y : 0.;    // diff_min
-        case 6: return _boxes[glyphid] ? _boxes[glyphid]->slant().tr.x : 0.;    // sum_max
-        case 7: return _boxes[glyphid] ? _boxes[glyphid]->slant().tr.y : 0.;    // diff_max
+        case 0: return (float)(glyph(glyphid)->theBBox().bl.x);                          // x_min
+        case 1: return (float)(glyph(glyphid)->theBBox().bl.y);                          // y_min
+        case 2: return (float)(glyph(glyphid)->theBBox().tr.x);                          // x_max
+        case 3: return (float)(glyph(glyphid)->theBBox().tr.y);                          // y_max
+        case 4: return (float)(_boxes[glyphid] ? _boxes[glyphid]->slant().bl.x : 0.);    // sum_min
+        case 5: return (float)(_boxes[glyphid] ? _boxes[glyphid]->slant().bl.y : 0.);    // diff_min
+        case 6: return (float)(_boxes[glyphid] ? _boxes[glyphid]->slant().tr.x : 0.);    // sum_max
+        case 7: return (float)(_boxes[glyphid] ? _boxes[glyphid]->slant().tr.y : 0.);    // diff_max
         default: return 0.;
     }
 }
